pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cvd-simulator"
version = "1.1.1"
description = "A comprehensive tool for simulating color vision deficiencies"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Michail Semoglou"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Processing",
]
requires-python = ">=3.10"
dependencies = [
    "daltonlens==0.1.5",
    "Pillow==12.1.1",
    "numpy==2.4.2",
    "tqdm>=4.66.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-benchmark>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.0.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
    "mypy>=1.10.0",
    "pydantic>=2.0.0",
    "psutil>=5.9.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.25.0",
    "setuptools>=61.0",
    "wheel>=0.37.0",
]

[project.scripts]
cvd-simulator = "cvd_simulator.interfaces.cli:main"

[project.urls]
Homepage = "https://github.com/MichailSemoglou/types-of-cvd-simulator-app"
Repository = "https://github.com/MichailSemoglou/types-of-cvd-simulator-app.git"
Issues = "https://github.com/MichailSemoglou/types-of-cvd-simulator-app/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src/cvd_simulator"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.pytest-benchmark]
min_rounds = 5
max_time = 10.0
warmup = true
sort = "mean"
group_by = "group"

[tool.hypothesis]
deadline = 5000
max_examples = 100
print_blob = true
phase.explicit = true
phase.reuse = true
phase.generate = true
phase.shrink = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".tox",
    ".eggs",
    "*.egg",
    "docs/conf.py",
]